requirements.txt
# FastAPI 和 Web框架
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.17

# 数据库
asyncpg>=0.30.0
sqlalchemy>=2.0.36
alembic>=1.14.0
pgvector>=0.3.6

# AI和向量
openai>=1.56.2
numpy>=1.24.0

# MCP相关
mcp>=1.1.0

# Threema加密
PyNaCl>=1.5.0
aiohttp>=3.11.10

# 配置和日志
pydantic>=2.10.3
pydantic-settings>=2.6.1
structlog>=24.4.0
python-dotenv>=1.0.1

# 测试（可选）
pytest>=8.3.4
pytest-asyncio>=0.25.0
httpx>=0.28.0
